Moving DOM references into components that need it.
